+++ packages/keyring-service/src/index.ts
@@ -3,11 +3,3 @@ export { KeyringService } from './keyring-service'
 
 // Keyring implementations
 export { SimpleKeyring, HdKeyring, KeystoneKeyring, ColdWalletKeyring } from './keyrings'
-
-// Storage adapters
-export { MemoryStorageAdapter } from './adapters/memory'
-export { ExtensionPersistStoreAdapter } from './adapters/extensionPersist'
-
-// Encryptor implementations
-export { BrowserPassworderEncryptor } from './encryptor/browser-encryptor'
-export { SimpleEncryptor } from './encryptor/simple-encryptor'