+++ packages/contact-book/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @unisat/contact-book
 
+## 0.1.8
+
+### Patch Changes
+
+- Updated dependencies
+  - @unisat/wallet-types@0.2.1
+
 ## 0.1.7
 
 ### Patch Changes