+++ packages/wallet-api/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @unisat/wallet-api
 
+## 0.1.10
+
+### Patch Changes
+
+- Updated dependencies
+  - @unisat/wallet-types@0.2.1
+
 ## 0.1.9
 
 ### Patch Changes