+++ packages/wallet-bitcoin/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @unisat/wallet-bitcoin
 
+## 0.2.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @unisat/wallet-types@0.2.1
+
 ## 0.2.0
 
 ### Minor Changes