+++ packages/wallet-types/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @unisat/wallet-types
 
+## 0.2.1
+
+### Patch Changes
+
+- fix
+
 ## 0.2.0
 
 ### Minor Changes